requirements.txt
SpeechRecognition
gTTS
google-generativeai
python-dotenv
streamlit
pymupdf
beautifulsoup4
requests

-e .

